requirements.txt
fastapi
openai
python-dotenv
sentence-transformers
scikit-learn